+++ src/Molecules/Tooltip/Tooltip.types.ts
@@ -1,18 +1,6 @@
 import React from 'react';
-import { InModal, Position } from 'common/PopOver/PopOver.types';
-import { BasePlacement, Rect } from '@popperjs/core';
+import { InModal, Position, OffsetArg } from 'common/PopOver/PopOver.types';
 import { Theme } from 'theme/theme.types';
-
-type OffsetFunctionArg = {
-  popper: Rect;
-  reference: Rect;
-  placement: BasePlacement;
-};
-
-type BareOffset = [number, number];
-type EmptyOffset = [];
-type Offset = BareOffset | EmptyOffset;
-export type OffsetArg = Offset | ((arg: OffsetFunctionArg) => Offset);
 
 export type Props = {
   label: React.ReactNode;
